# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory-API",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://jeremybecker.ch/",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/book": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "List books",
                "description": "Get a List of all books.",
                "operationId": "listBook",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Create book",
                "description": "Create a new book.",
                "operationId": "storeBook",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The title of the book.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The id of the user created the book.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location_id",
                        "in": "path",
                        "description": "The location the book is stored in.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "isbn",
                        "in": "path",
                        "description": "The isbn number of the book.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "path",
                        "description": "The author of the book.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "excerpt",
                        "in": "path",
                        "description": "A short text that describes the content of the book.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "release_date",
                        "in": "path",
                        "description": "The date the book was released on.",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "language",
                        "in": "path",
                        "description": "The language the book is written in.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/book/{id}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Show book",
                "description": "Display a book",
                "operationId": "showBook",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Delete book",
                "description": "Delete a book.",
                "operationId": "deleteBook",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Book"
                ],
                "summary": "Update book",
                "description": "Update a book.",
                "operationId": "updateBook",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The title of the book.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The id of the user created the book.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location_id",
                        "in": "path",
                        "description": "The location the book is stored in.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "isbn",
                        "in": "path",
                        "description": "The isbn number of the book.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "path",
                        "description": "The author of the book.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "excerpt",
                        "in": "path",
                        "description": "A short text that describes the content of the book.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "release_date",
                        "in": "path",
                        "description": "The date the book was released on.",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "language",
                        "in": "path",
                        "description": "The language the book is written in.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    }
}